Доделал ремиж -перейти в положение
Добавил новое условие останова для этого режима. Исправил проверку
ограничений на задаваемую позицию. вывожу в дебаг попытки задать большее
значение чем положено. Небольшой рефакторинг.

+++ CommandController.h
@@ -15,7 +15,7 @@
 #include <math.h>
 #include <cmath>
 #include "BufferController.h"
-#include "AR600ControllerConf.h"
+#include "ConfigController.h"
 #include <QStateMachine>
 #include <QState>
 
@@ -41,9 +41,9 @@ private:
     int mTimeToGo;
     int mDestPos;
     int mStartPos;
-    int mCurrentPos;
+    double mCurrentPos;
     int mCurrentTime;
-    int mStepPos;
+    double mStepPos;
     int mDriverNumberBuffer;
     bool IsGoToPosState;//состояние выполнения перехода в позицию за время
